Fixes #365 (#388)
 Use u64 to represent block heights in light wallet CompactBlock #388

+++ pd/src/state/reader.rs
@@ -425,7 +425,7 @@ impl Reader {
 
             for height in start_height..=end_height {
                 let mut compact_block = CompactBlock {
-                    height: height as u32,
+                    height: height as u64,
                     fragments: vec![],
                     nullifiers: vec![],
                 };
